Tray icon and auto hide on startup
Added a tray icon that allows the application to be hidden on startup.
Added an option in optional settings that allows the user to select whether the application should be hidden on start. Default value is value.
The application window will be shown until the initialisation is complete. This allows the application not to hide if any errors occurred during initialisation.